+++ src/parse-config.ts
@@ -11,13 +11,18 @@ export type CookieConnoisseurConfig = {
         [url: string]: {
             path: string;
         };
-    }
+    },
+    localDirectories: {
+        [urlPrefix: string]: {
+            path: string;
+        };
+    },
 }
 
 export const configPath = 'cookie-connoisseur.config.json';
 
 export async function parseConfigFile(): Promise<CookieConnoisseurConfig> {
-    let config: CookieConnoisseurConfig = {customURLs: {}, localFiles: {}};
+    let config: CookieConnoisseurConfig = {customURLs: {}, localFiles: {}, localDirectories: {}};
     if(!existsSync(configPath)) {
         return config;
     }
@@ -26,7 +31,7 @@ export async function parseConfigFile(): Promise<CookieConnoisseurConfig> {
     if(!content || typeof content != "object")
         return config;
 
-    if('customURLs' in content && typeof content['customURLs'] == "object") {
+    if('customURLs' in content && Array.isArray(content['customURLs'])) {
         for(let obj of content.customURLs) {
             if('url' in obj && typeof obj.url == 'string') {
                 config.customURLs[obj.url] = {};
@@ -34,11 +39,20 @@ export async function parseConfigFile(): Promise<CookieConnoisseurConfig> {
         }
     }
 
-    if('localFiles' in content && typeof content['localFiles'] == "object") {
+    if('localFiles' in content && Array.isArray(content['localFiles'])) {
         for(let obj of content.localFiles) {
             if('url' in obj && typeof obj.url == 'string') {
                 if('path' in obj && typeof obj.path == 'string')
                     config.localFiles[obj.url] = {path: obj.path};
+            }
+        }
+    }
+
+    if('localDirectories' in content && Array.isArray(content['localDirectories'])) {
+        for(let obj of content.localDirectories) {
+            if('url' in obj && typeof obj.url == 'string') {
+                if('path' in obj && typeof obj.path == 'string')
+                    config.localDirectories[obj.url] = {path: obj.path};
             }
         }
     }